// apps/backend/src/types/database.ts
export type Database = {
  public: {
    Tables: {
      User: {
        Row: {
          id: string;
          email: string;
          firstName: string;
          lastName: string;
          phone: string | null;
          isDeleted: boolean;
          createdAt: string;
          updatedAt: string;
          organizationId: string | null;
          accessLevel: string | null;
          isSupportUser: boolean;
          hasSeenProductTour: boolean;
          groupView: string;
          photoView: string;
          savedDashboardView: string;
          onboardingStatus: any | null;
          productTourData: any | null;
          inviteId: string | null;
          token: string | null;
          removed: boolean;
        };
        Insert: {
          id: string;
          email: string;
          firstName?: string;
          lastName?: string;
          phone?: string | null;
          isDeleted?: boolean;
          createdAt?: string;
          updatedAt?: string;
          organizationId?: string | null;
          accessLevel?: string | null;
          isSupportUser?: boolean;
          hasSeenProductTour?: boolean;
          groupView?: string;
          photoView?: string;
          savedDashboardView?: string;
          onboardingStatus?: any | null;
          productTourData?: any | null;
          inviteId?: string | null;
          token?: string | null;
          removed?: boolean;
        };
        Update: {
          id?: string;
          email?: string;
          firstName?: string;
          lastName?: string;
          phone?: string | null;
          isDeleted?: boolean;
          createdAt?: string;
          updatedAt?: string;
          organizationId?: string | null;
          accessLevel?: string | null;
          isSupportUser?: boolean;
          hasSeenProductTour?: boolean;
          groupView?: string;
          photoView?: string;
          savedDashboardView?: string;
          onboardingStatus?: any | null;
          productTourData?: any | null;
          inviteId?: string | null;
          token?: string | null;
          removed?: boolean;
        };
      };
      Organization: {
        Row: {
          id: number;
          name: string;
          phoneNumber: string | null;
          address: string | null;
          faxNumber: string | null;
          size: string;
          isDeleted: boolean;
          logoId: string | null;
          lat: number | null;
          lng: number | null;
          subscriptionId: string | null;
          subscriptionPlan: string | null;
          customerId: string | null;
          maxUsersForSubscription: number | null;
          freeTrialEndsAt: string | null;
          subscriptionStatus: string | null;
          createdAt: string;
          updatedAt: string;
          publicId: string;
          owner: string | null;
          stripeSessionId: string | null;
        };
        Insert: {
          id?: number;
          name: string;
          phoneNumber?: string | null;
          address?: string | null;
          faxNumber?: string | null;
          size: string;
          isDeleted?: boolean;
          logoId?: string | null;
          lat?: number | null;
          lng?: number | null;
          subscriptionId?: string | null;
          subscriptionPlan?: string | null;
          customerId?: string | null;
          maxUsersForSubscription?: number | null;
          freeTrialEndsAt?: string | null;
          subscriptionStatus?: string | null;
          createdAt?: string;
          updatedAt?: string;
          publicId: string;
          owner?: string | null;
          stripeSessionId?: string | null;
        };
        Update: {
          id?: number;
          name?: string;
          phoneNumber?: string | null;
          address?: string | null;
          faxNumber?: string | null;
          size?: string;
          isDeleted?: boolean;
          logoId?: string | null;
          lat?: number | null;
          lng?: number | null;
          subscriptionId?: string | null;
          subscriptionPlan?: string | null;
          customerId?: string | null;
          maxUsersForSubscription?: number | null;
          freeTrialEndsAt?: string | null;
          subscriptionStatus?: string | null;
          createdAt?: string;
          updatedAt?: string;
          publicId?: string;
          owner?: string | null;
          stripeSessionId?: string | null;
        };
      };
      UserToOrganization: {
        Row: {
          id: number;
          userId: string;
          organizationId: number;
          role: string | null;
          isAdmin: boolean;
          isDeleted: boolean;
          createdAt: string;
          updatedAt: string;
          accessLevel: string | null;
        };
        Insert: {
          id?: number;
          userId: string;
          organizationId: number;
          role?: string | null;
          isAdmin?: boolean;
          isDeleted?: boolean;
          createdAt?: string;
          updatedAt?: string;
          accessLevel?: string | null;
        };
        Update: {
          id?: number;
          userId?: string;
          organizationId?: number;
          role?: string | null;
          isAdmin?: boolean;
          isDeleted?: boolean;
          createdAt?: string;
          updatedAt?: string;
          accessLevel?: string | null;
        };
      };
    };
  };
};